# packages/api-client/openapi.json
{
...
  "info": {
    "contact": {},
    "description": "REST API for the Hà Tĩnh immersive tourism platform foundation.",
    "title": "Hà Tĩnh Immersive API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/destinations/{slug}/immersive-manifest": {
      "get": {
        "operationId": "getImmersiveManifest",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "defaultSceneId": {
                      "format": "uuid",
                      "nullable": true,
                      "type": "string"
                    },
                    "destination": {
                      "properties": {
                        "categoryId": {
                          "format": "uuid",
                          "nullable": true,
                          "type": "string"
                        },
                        "categoryLabel": {
                          "nullable": true,
                          "type": "string"
                        },
                        "coverImageUrl": {
                          "nullable": true,
                          "type": "string"
                        },
                        "coverMediaId": {
                          "format": "uuid",
                          "nullable": true,
                          "type": "string"
                        },
                        "defaultSceneId": {
                          "format": "uuid",
                          "nullable": true,
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "geoPoint": {
                          "nullable": true,
                          "properties": {
                            "latitude": {
                              "example": 18.3421,
                              "type": "number"
                            },
                            "longitude": {
                              "example": 105.9032,
                              "type": "number"
                            }
                          },
                          "required": ["latitude", "longitude"],
                          "type": "object"
                        },
                        "id": {
                          "format": "uuid",
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "slug": {
                          "type": "string"
                        },
                        "status": {
                          "enum": ["draft", "published", "archived"],
                          "type": "string"
                        },
                        "summary": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "slug",
                        "name",
                        "summary",
                        "coverImageUrl",
                        "categoryLabel",
                        "status",
                        "description",
                        "categoryId",
                        "defaultSceneId",
                        "geoPoint",
                        "coverMediaId"
                      ],
                      "type": "object"
                    },
                    "hotspots": {
                      "items": {
                        "properties": {
                          "id": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "payload": {
                            "additionalProperties": true,
                            "type": "object"
                          },
                          "pitch": {
                            "maximum": 90,
                            "minimum": -90,
                            "type": "number"
                          },
                          "sceneId": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "status": {
                            "enum": ["draft", "published", "archived"],
                            "type": "string"
                          },
                          "type": {
                            "enum": ["information", "media", "audio", "external"],
                            "type": "string"
                          },
                          "yaw": {
                            "exclusiveMaximum": true,
                            "maximum": 360,
                            "minimum": 0,
                            "type": "number"
                          }
                        },
                        "required": ["id", "sceneId", "type", "yaw", "pitch", "payload", "status"],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "links": {
                      "items": {
                        "properties": {
                          "bidirectional": {
                            "type": "boolean"
                          },
                          "fromSceneId": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "id": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "pitch": {
                            "maximum": 90,
                            "minimum": -90,
                            "type": "number"
                          },
                          "sortOrder": {
                            "type": "integer"
                          },
                          "toSceneId": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "yaw": {
                            "exclusiveMaximum": true,
                            "maximum": 360,
                            "minimum": 0,
                            "type": "number"
                          }
                        },
                        "required": [
                          "id",
                          "fromSceneId",
                          "toSceneId",
                          "yaw",
                          "pitch",
                          "bidirectional",
                          "sortOrder"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "nodes": {
                      "items": {
                        "properties": {
                          "altitude": {
                            "nullable": true,
                            "type": "number"
                          },
                          "destinationId": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "id": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "initialFov": {
                            "type": "number"
                          },
                          "initialHeading": {
                            "type": "number"
                          },
                          "initialPitch": {
                            "type": "number"
                          },
                          "lat": {
                            "type": "number"
                          },
                          "lng": {
                            "type": "number"
                          },
                          "name": {
                            "type": "string"
                          },
                          "panoramaAssetId": {
                            "format": "uuid",
                            "nullable": true,
                            "type": "string"
                          },
                          "panoramaAssetStatus": {
                            "enum": ["pending", "uploaded", "processing", "ready", "failed"],
                            "nullable": true,
                            "type": "string"
                          },
                          "sortOrder": {
                            "type": "integer"
                          },
                          "status": {
                            "enum": ["draft", "published", "archived"],
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "destinationId",
                          "name",
                          "lat",
                          "lng",
                          "altitude",
                          "panoramaAssetId",
                          "panoramaAssetStatus",
                          "initialHeading",
                          "initialPitch",
                          "initialFov",
                          "status",
                          "sortOrder"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "required": ["destination", "defaultSceneId", "nodes", "links", "hotspots"],
                  "type": "object"
                }
              }
            },
            "description": "Published immersive scene graph manifest."
          }
        },
        "summary": "",
        "tags": ["immersive"]
      }
    },
...
    "/api/v1/scenes/{sceneId}": {
      "get": {
        "operationId": "getScene",
        "parameters": [
          {
            "in": "path",
            "name": "sceneId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "altitude": {
                      "nullable": true,
                      "type": "number"
                    },
                    "destinationId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "initialFov": {
                      "type": "number"
                    },
                    "initialHeading": {
                      "type": "number"
                    },
                    "initialPitch": {
                      "type": "number"
                    },
                    "lat": {
                      "type": "number"
                    },
                    "lng": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "panoramaAssetId": {
                      "format": "uuid",
                      "nullable": true,
                      "type": "string"
                    },
                    "panoramaAssetStatus": {
                      "enum": ["pending", "uploaded", "processing", "ready", "failed"],
                      "nullable": true,
                      "type": "string"
                    },
                    "sortOrder": {
                      "type": "integer"
                    },
                    "status": {
                      "enum": ["draft", "published", "archived"],
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "destinationId",
                    "name",
                    "lat",
                    "lng",
                    "altitude",
                    "panoramaAssetId",
                    "panoramaAssetStatus",
                    "initialHeading",
                    "initialPitch",
                    "initialFov",
                    "status",
                    "sortOrder"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Published immersive scene node."
          }
        },
        "summary": "",
        "tags": ["immersive"]
      }
    },
    "/api/v1/scenes/{sceneId}/neighbors": {
      "get": {
        "operationId": "getSceneNeighbors",
        "parameters": [
          {
            "in": "path",
            "name": "sceneId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "properties": {
                      "link": {
                        "properties": {
                          "bidirectional": {
                            "type": "boolean"
                          },
                          "fromSceneId": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "id": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "pitch": {
                            "maximum": 90,
                            "minimum": -90,
                            "type": "number"
                          },
                          "sortOrder": {
                            "type": "integer"
                          },
                          "toSceneId": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "yaw": {
                            "exclusiveMaximum": true,
                            "maximum": 360,
                            "minimum": 0,
                            "type": "number"
                          }
                        },
                        "required": [
                          "id",
                          "fromSceneId",
                          "toSceneId",
                          "yaw",
                          "pitch",
                          "bidirectional",
                          "sortOrder"
                        ],
                        "type": "object"
                      },
                      "scene": {
                        "properties": {
                          "altitude": {
                            "nullable": true,
                            "type": "number"
                          },
                          "destinationId": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "id": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "initialFov": {
                            "type": "number"
                          },
                          "initialHeading": {
                            "type": "number"
                          },
                          "initialPitch": {
                            "type": "number"
                          },
                          "lat": {
                            "type": "number"
                          },
                          "lng": {
                            "type": "number"
                          },
                          "name": {
                            "type": "string"
                          },
                          "panoramaAssetId": {
                            "format": "uuid",
                            "nullable": true,
                            "type": "string"
                          },
                          "panoramaAssetStatus": {
                            "enum": ["pending", "uploaded", "processing", "ready", "failed"],
                            "nullable": true,
                            "type": "string"
                          },
                          "sortOrder": {
                            "type": "integer"
                          },
                          "status": {
                            "enum": ["draft", "published", "archived"],
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "destinationId",
                          "name",
                          "lat",
                          "lng",
                          "altitude",
                          "panoramaAssetId",
                          "panoramaAssetStatus",
                          "initialHeading",
                          "initialPitch",
                          "initialFov",
                          "status",
                          "sortOrder"
                        ],
                        "type": "object"
                      }
                    },
                    "required": ["link", "scene"],
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Published neighboring scene nodes."
          }
        },
        "summary": "",
        "tags": ["immersive"]
      }
    }
  },
  "servers": [],
...
}